+++ frontend/app/src/data/factories.ts
@@ -1,6 +1,10 @@
 import { currencies } from '@/data/currencies';
 import { Defaults } from '@/data/defaults';
-import { AccountingSettings, GeneralSettings } from '@/types/user';
+import {
+  AccountingSettings,
+  CostBasisMethod,
+  GeneralSettings
+} from '@/types/user';
 
 export const defaultGeneralSettings = (): GeneralSettings => ({
   uiFloatingPrecision: Defaults.FLOATING_PRECISION,
@@ -28,5 +32,6 @@ export const defaultAccountingSettings = (): AccountingSettings => ({
   taxfreeAfterPeriod: null,
   accountForAssetsMovements: true,
   calculatePastCostBasis: true,
-  taxableLedgerActions: []
+  taxableLedgerActions: [],
+  costBasisMethod: CostBasisMethod.Fifo
 });